Fix AI using 🔍 and shooting itself
useItem('🔍') will also automatically shoot the gun. On a blank round,
the AI is intended to point the gun towards itself, which works fine. If
a live round is found, the AI is intended to point the gun towards the
player, optionally doubling damage beforehand.
Since `effector` has not been set previously, the AI would shoot itself
with a live round though. This is fixed by passing the player as
`effector` when the AI uses 🔍.